% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{blank_to_na}
\alias{blank_to_na}
\title{Convert all blank to na in dataset}
\usage{
blank_to_na(d)
}
\arguments{
\item{d}{a dataset from cgdsr}
}
\value{
dataset complete
}
\description{
This function converts all blank values in the dataset to NA,
this is very useful in real world datasets that contain masked NA values
}
