% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy.R
\name{brier_score}
\alias{brier_score}
\title{Calculate Integrated Brier Score}
\usage{
brier_score(pred_newdata, test_data, cutoff = NA)
}
\arguments{
\item{pred_newdata}{a dataset prediction made by the training set (or CV fold)  \cr
tet_data the hold out dataset \cr
cutoff an optional parameter to integrate the Brier score in a different time, default is max observed time \cr}
}
\value{
pp_bried the posterior predicted integrated brier score
}
\description{
This function recieves a training set and a test set and calculates
different measures of accuracy and loss functions
}
