% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical.R
\name{categories_to_index}
\alias{categories_to_index}
\title{Coerce categorical variables in index}
\usage{
categories_to_index(d, ...)
}
\arguments{
\item{d}{a dataset \cr
... vars to be coerced \cr}
}
\value{
d a dataset with coerced covariates
}
\description{
This function allows to coerce categorical variables into index, \cr
}
