% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clinicaldata}
\alias{clinicaldata}
\title{Clinical data from the TCGA dataset}
\format{A data frame with230  rows and  19 variables:
\describe{
  \item{patient_id}{Randomly completely Anonamysed patient ID unique for each row}
  \item{cancer_type_detailed}{A longer explanation of cancer subtype}
  ...
}}
\source{
\url{https://cancergenome.nih.gov/}
}
\usage{
clinicaldata
}
\description{
A dataset containing the survival outcome in lung cancer TCGA trial
}
\keyword{datasets}
