% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_training_test_split.R
\name{create_training_test_set}
\alias{create_training_test_set}
\title{Create training test split}
\usage{
create_training_test_set(d, y = censoring.status, seed = 111,
  percent = 0.8)
}
\arguments{
\item{d}{a dataset \cr
seed  \cr
percent the percentage of data that goes to training}
}
\value{
a list with train and test set
}
\description{
This function plots the TTE distribution
}
