% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{drop_na}
\alias{drop_na}
\title{Clean dataset}
\usage{
drop_na(d, ..., time = os_months, n = NA)
}
\arguments{
\item{d}{a dataset \cr
... vars to be evaluated \cr
n optional parameter to assure number of rows deleted, default NA}
}
\value{
d a clean dataset
}
\description{
This function is only to be used if there is the will to delete missing data.
All rows with missing data from the vars passed will be deleted
}
