% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_features.R
\name{extract_pars_cv}
\alias{extract_pars_cv}
\title{Extract particular features from cv output}
\usage{
extract_pars_cv(cv_obj, feature, model, time = os_months,
  status = os_deceased, ...)
}
\arguments{
\item{data}{: cv_obj should be a list \cr
feature: brier_pred, roc_pred...}
}
\value{
a coxph fit object
}
\description{
This function should not be called directly
}
