% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_poisson_trick.R
\name{fit_poisson}
\alias{fit_poisson}
\title{Fit a Poisson distribution}
\usage{
fit_poisson(d, t_start = tstart, t_stop = time2event, t_dur = t_dur,
  status = os_event, formula = tobacco_smoking_history_indicator)
}
\arguments{
\item{d}{a dataset \cr
formula covariate matrix \cr}
}
\value{
a fit objects ( smoth splines).
}
\description{
This function plots the TTE distribution,   Important aspects of the modelling approach: Dependent variable is the binary event indicator.  Adding an intercept is recommendable to provide a decent and interpretable reference category for factor variables.
}
