% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predsurv.R
\name{fit_stan_surv}
\alias{fit_stan_surv}
\title{Fit stan survival model
Fit stan survival model}
\usage{
fit_stan_surv(dat, g, x_vars, time_var = "time", status_var = "status",
  method = "L1", p0 = 5, nv_max = 40, save_file, ...)
}
\arguments{
\item{g}{additional data frame for bootstrapping}

\item{x_vars}{vars to analysis}

\item{time_var}{defults to time}

\item{status_var}{defaults to status}

\item{p0}{prior guess for the number of relevant variables}

\item{save_file}{should results be saved on file}

\item{d_train}{dataframe}

\item{ncores}{for parallel.}

\item{alpha}{alpha for elastic net}

\item{relaxed}{apply relaxed lasso set to TRUE}
}
\description{
Fit stan survival model
Fit stan survival model
}
