% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploration.R
\name{fit_weibull}
\alias{fit_weibull}
\alias{fit_weibull}
\title{Fit Exponential distribution}
\usage{
fit_weibull(d)

fit_weibull(d)
}
\arguments{
\item{d}{a dataset \cr
x customise plot title \cr
event_type codification of event type , default DECEASED}

\item{d}{a dataset \cr
x customise plot title \cr
event_type codification of event type , default DECEASED}
}
\value{
d a clean dataset

d a clean dataset
}
\description{
This function fits the Weibull distribution

This function fits the Weibull distribution
}
