% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predsurv.R
\name{forward_sel}
\alias{forward_sel}
\title{Fit forward stepwise selection survival model
Get the cost and coefficients for forward search survival analysis.}
\usage{
forward_sel(d_train, g, x_vars, time_var = "time",
  status_var = "status", penalty = "AIC", save_file)
}
\arguments{
\item{d_train}{dataframe}

\item{g}{additional data frame for bootstrapping}

\item{x_vars}{vars to analysis}

\item{time_var}{defults to time}

\item{status_var}{defaults to status}

\item{penalty}{information criteria to use: AIC or BIC}

\item{save_file}{should results be saved on file}

\item{ncores}{for parallel.}
}
\description{
Fit forward stepwise selection survival model
Get the cost and coefficients for forward search survival analysis.
}
