% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cross-validation.R
\name{fun_cv}
\alias{fun_cv}
\title{Compute C-V iBrier}
\usage{
fun_cv(data, method = "forward", time = os_months, status = os_deceased,
  K = 10)
}
\arguments{
\item{obj}{: survival model object for example a glmnet fit. \cr
coef: coefficient beta is the default. \cr
new_data: a test holdout dataframe default test\cr
train_data : train dataframe default train\cr}
}
\value{
a coxph fit object
}
\description{
Predict integration Brier Score in a new dataset.
}
