% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_surv.R
\name{fun_fit}
\alias{fun_fit}
\title{Fit Forward Selection}
\usage{
fun_fit(train, time = os_months, status = os_deceased, method = "forward",
  fit, penalty = "BIC", tree_control = "Univariate")
}
\arguments{
\item{train}{: train data \cr
time : time variable \cr
status : status variable \cr
method: forward, backward or both \cr
fit : fit model: "stepwise", "random forest"}
}
\value{
a coxph forward stepwise fit object
}
\description{
Fit the Forward Selection method
}
