% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_test.R
\name{fun_mp}
\alias{fun_mp}
\title{Measure Performance}
\usage{
fun_mp(obj, test_data, fit = "Lasso", subject = subject, time = os_months,
  status = os_deceased, event_type = 1, pred = "Brier", all = FALSE,
  integrated = TRUE, reference = FALSE, noboot = 0, mc = FALSE, ...)
}
\arguments{
\item{obj}{: survival model object for example a glmnet fit. \cr
coef: coefficient beta is the default. \cr
new_data: a test holdout dataframe default test\cr
train_data : train dataframe default train\cr
pred : prediction error Brier, ROC or C-Index
adapted : in Lasso allows to swap to "adapated Lasso", otherwise will take usual Lasso \cr}
}
\value{
a coxph fit object
}
\description{
This function takes a survival fit and following the Cox model (or random forest), estimates the hazard for individual from the model coefficient, while the baseline hazard is estimated with the Nelson-Aalen method. Gives prediction error measures Brier, c index, loglik, auc
}
