% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_surv.R
\name{fun_score}
\alias{fun_score}
\title{Fit and measure measures of prognostic performance
Prediction of various scores}
\usage{
fun_score(train, fit, data, prediction)
}
\arguments{
\item{obj}{: survival model object for example a glmnet fit. \cr
coef: coefficient beta is the default. \cr
new_data: a test holdout dataframe default test\cr
train_data : train dataframe default train\cr}
}
\value{
a coxph fit object
}
\description{
Fit and measure measures of prognostic performance
Prediction of various scores
}
