% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_surv.R
\name{fun_train}
\alias{fun_train}
\title{Fit Model to Training data}
\usage{
fun_train(train, time = os_months, status = os_deceased, fit,
  penalty = "BIC")
}
\arguments{
\item{train}{: train data \cr
time : time variable \cr
status : status variable \cr
method: forward, backward or both \cr
fit : fit model: "Stepwise", "Random forest"}
}
\value{
a coxph forward Stepwise fit object
}
\description{
Train data to specific feature selection method
}
