% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_surv2.R
\name{fun_train2}
\alias{fun_train2}
\title{Fit Model to Training data}
\usage{
fun_train2(hold_out, data, time = os_months, status = os_deceased, fit,
  subject = patient_id, lambda = 0.001, preCut = FALSE, ...)
}
\arguments{
\item{train}{: train data \cr
time : time variable \cr
status : status variable \cr
method: forward, backward or both \cr
fit : fit model: "Stepwise", "Random forest"}
}
\value{
a coxph forward Stepwise fit object
}
\description{
Train data to specific feature selection method
}
