% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gbm}
\alias{gbm}
\title{GBM data}
\format{A list object containing three data matrices: copy number, methylation and mRNA expression in 55 samples.
\describe{
  \item{time}{Actual survival months observed}
  \item{cens}{censoring indicator 1 observed, 0 censored}
  \item{featurenames}{feature names in the gene expression matrix}
  ...
}}
\usage{
gbm
}
\description{
This is a subset of the glioblastoma dataset from the cancer genome atlas (TCGA) GBM study (2009) used in Shen et al. (2012).
}
\references{
Ronglai Shen, Qianxing Mo, Nikolaus Schultz, Venkatraman E. Seshan, Adam B. Olshen, Jason Huse, Marc Ladanyi, Chris Sander. (2012). Integrative Subtype Discovery in Glioblastoma Using iCluster. PLoS ONE 7, e35236
}
\keyword{datasets}
