% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_stan_data.R
\name{gen_long_data}
\alias{gen_long_data}
\title{This function generates the long data neeeded for a piecewise exponential model}
\usage{
gen_long_data(d, time = os_months, status = os_status,
  sample_id = patient_id, k = seq_along(tau), event_type = "DECEASED")
}
\arguments{
\item{d}{a dataset \cr
formula a formula to pass covariates \cr
k the number of intervals that want to be done
c an optional parameter to specify lenghts of intervals}
}
\value{
a long format (count model) dataset
}
\description{
This function generates the long data neeeded for a piecewise exponential model
}
