% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_stan_data.R
\name{gen_stan_data}
\alias{gen_stan_data}
\title{This function generates the stan data for a clinical PEM model, the rason for not #' introducing genomic covariates is that it takes too much memory, for instance the #' genomic covariates will be handled ina module in Python.}
\usage{
gen_stan_data(d, formula = as.formula(~1), time = t_dur,
  status = os_event, sample_id = patient_id, time_id = t_id)
}
\arguments{
\item{d}{a dataset \cr
formula a formula to pass covariates \cr
time = time to event
status = status covariate
sample_id = unique id for each patient
time_id = unique id for each observation}
}
\value{
a list for stan
}
\description{
This function generates the stan data for a clinical PEM model, the rason for not #' introducing genomic covariates is that it takes too much memory, for instance the #' genomic covariates will be handled ina module in Python.
}
