% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\name{get_data}
\alias{get_data}
\title{Get high dimensional survival data.}
\usage{
get_data(dataset = "survival")
}
\arguments{
\item{rate}{rate parameter for the exponential distribution \cr
beta beta regression coefficient\cr}
}
\value{
simulated data
}
\description{
This function gets a survival data experiment
}
