% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute_gene_matrix}
\alias{impute_gene_matrix}
\title{Data Impute}
\usage{
impute_gene_matrix(genedata)
}
\arguments{
\item{genedata:}{genomic dataset \cr}
}
\value{
a full dataset
}
\description{
This function imputes a gene matrix, currently with knn
}
