% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predsurv.R
\name{lasso_sel}
\alias{lasso_sel}
\title{Fit elastic net selection survival model
Get the cost and coefficients for elastic net survival analysis.}
\usage{
lasso_sel(d_train, g, x_vars, time_var = "time", status_var = "status",
  alpha = 1, relaxed = F, ncores, save_file)
}
\arguments{
\item{d_train}{dataframe}

\item{g}{additional data frame for bootstrapping}

\item{x_vars}{vars to analysis}

\item{time_var}{defults to time}

\item{status_var}{defaults to status}

\item{alpha}{alpha for elastic net}

\item{relaxed}{apply relaxed lasso set to TRUE}

\item{ncores}{for parallel.}

\item{save_file}{should results be saved on file}

\item{correction}{correction to apply.}
}
\description{
Fit elastic net selection survival model
Get the cost and coefficients for elastic net survival analysis.
}
