% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{LoadToEnvironment}
\alias{LoadToEnvironment}
\title{Load Data to environment}
\usage{
LoadToEnvironment(RData, env = new.env())
}
\arguments{
\item{rdata}{\cr}
}
\value{
a coxph fit object
}
\description{
Load Data to environment
}
