% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metabric_clinical_data}
\alias{metabric_clinical_data}
\title{Metabric clinical data}
\format{A datset with 3 covariates and 2 survival variables.
\describe{
  \item{os_months}{Actual survival months observed}
  \item{os_status}{censoring indicator 1 observed, 0 censored}
  ...
}}
\usage{
metabric_clinical_data
}
\description{
This dataset includes the clinical data for the study METABRIC.
}
\references{
Curtis C, Shah SP, Chin SF, Turashvili G, Rueda OM, Dunning MJ, Speed D, Lynch AG, Samarajiwa S, Yuan Y, et al. The genomic and transcriptomic architecture of 2,000 breast tumours reveals novel subgroups. Nature. 2012; 486:346–352. [PubMed: 22522925]
}
\keyword{datasets}
