% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_surv.R
\name{mod_pred}
\alias{mod_pred}
\title{Cox Model Prediction}
\usage{
mod_pred(obj, time = os_months, status = os_deceased, beta = beta,
  test_data = test, train_data = train)
}
\arguments{
\item{obj}{: survival model object for example a glmnet fit. \cr
coef: coefficient beta is the default. \cr
new_data: a test holdout dataframe default test\cr
train_data : train dataframe default train\cr}
}
\value{
a coxph fit object
}
\description{
This function takes a survival fit and following the Cox model, estimates the hazard for individual from the model coefficient, while the baseline hazard is estimated with the Nelson-Aalen method.
}
