% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PenalisedLikelihood.R
\name{penalised_likelihood_fit}
\alias{penalised_likelihood_fit}
\title{Calculate Penalised Likelihood}
\usage{
penalised_likelihood_fit(formula = as.formula(~1), y, family = "cox",
  plot = FALSE)
}
\arguments{
\item{train}{datset for training  \cr
formula \cr
cutoff an optional parameter to integrate the Brier score in a different time, default is max observed time \cr}
}
\value{
pp_bried the posterior predicted integrated brier score
}
\description{
This function calculates the penalised likelihood of the Cox model
}
