% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pred.R
\name{plot_brier}
\alias{plot_brier}
\title{Plot Brier Score}
\usage{
plot_brier(brier_score)
}
\arguments{
\item{obj}{: survival model object for example a glmnet fit. \cr
coef: coefficient beta is the default. \cr
new_data: a test holdout dataframe default test\cr
train_data : train dataframe default train\cr}
}
\value{
a coxph fit object
}
\description{
Plot Brier Score
}
