% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pred.R
\name{plot_brier2}
\alias{plot_brier2}
\title{Plot Brier Score}
\usage{
plot_brier2(..., x.var = "Time", y.var = "Brier Score", x.lab = x.var,
  y.lab = y.var, ident = FALSE)
}
\arguments{
\item{obj}{: survival model object for example a glmnet fit. \cr
coef: coefficient beta is the default. \cr
new_data: a test holdout dataframe default test\cr
train_data : train dataframe default train\cr}
}
\value{
a coxph fit object
}
\description{
Plot Brier Score
}
