% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploration.R
\name{plot_correlation}
\alias{plot_correlation}
\title{The function for plot correlation between explanatory variables}
\usage{
plot_correlation(d, time = os_months, status = os_status,
  event_type = "DECEASED", x = "survival")
}
\arguments{
\item{d}{a dataset \cr
x customise plot title \cr
event_type codification of event type , default DECEASED}
}
\value{
d a clean dataset
}
\description{
The function for plot correlation between explanatory variables
}
