% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_surv.R
\name{plot_cox_pred}
\alias{plot_cox_pred}
\title{Plot Cox Model Prediction}
\usage{
plot_cox_pred(train_data = train, time = os_months, status = os_deceased,
  ...)
}
\arguments{
\item{data}{: survival dataset \cr
... : covariates to evaluate}
}
\value{
a coxph fit object
}
\description{
This function takes a survival fit and plots, right now only works with factor
}
\examples{
http://staff.pubhealth.ku.dk/~tag/Teaching/share/R-tutorials/SurvivalAnalysis.html
}
