% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{plot_na}
\alias{plot_na}
\title{Plot all na
This function gives an explanatory plot of NA values using VIM package}
\usage{
plot_na(d)
}
\arguments{
\item{d}{a dataset from cgdsr}
}
\value{
dataset complete
}
\description{
Plot all na
This function gives an explanatory plot of NA values using VIM package
}
