% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_poisson_trick.R
\name{plot_pred_poisson}
\alias{plot_pred_poisson}
\title{Prediction plot for the PEM}
\usage{
plot_pred_poisson(d, time = os_months, status = os_status,
  event_type = "DECEASED", plotdata = plotframe, predtime = Time,
  predsurv = S1, formula = as.formula(~1))
}
\arguments{
\item{d}{a dataset \cr
x customise plot title \cr
event_type codification of event type , default DECEASED}
}
\value{
d a clean dataset
}
\description{
This function plots the actual Kaplan and Meier and overlies the PEM prediction, follows the previous predict function
}
