% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploration.R
\name{plot_tte_dist}
\alias{plot_tte_dist}
\title{Time to Event distribution}
\usage{
plot_tte_dist(d, time = os_months, status = os_status, alpha = 0.5)
}
\arguments{
\item{d}{a dataset \cr
... vars to be evaluated \cr
n optional parameter to assure number of rows deleted, default NA}
}
\value{
d a clean dataset
}
\description{
This function plots the TTE distribution
}
