% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_surv.R
\name{pred_error}
\alias{pred_error}
\title{Cox Model Prediction}
\usage{
pred_error(obj, train_data = train, test_data = test, time = os,
  status = status, event_type = 1, pred = "Brier", fit = "stepwise",
  adapted = "default")
}
\arguments{
\item{obj}{: survival model object for example a glmnet fit. \cr
coef: coefficient beta is the default. \cr
new_data: a test holdout dataframe default test\cr
train_data : train dataframe default train\cr
pred : prediction error Brier, ROC or C-Index
adapted : in Lasso allows to swap to "adapated Lasso", otherwise will take usual lasso \cr}
}
\value{
a coxph fit object
}
\description{
This function takes a survival fit and following the Cox model, estimates the hazard for individual from the model coefficient, while the baseline hazard is estimated with the Nelson-Aalen method.
}
