% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_poisson_trick.R
\name{pred_poisson}
\alias{pred_poisson}
\title{Make predictions for the PEM. Series of functions in the PEM MLE.}
\usage{
pred_poisson(d, time = os_months, status = os_event, event_type = 1,
  k = seq_along(tau))
}
\arguments{
\item{d}{a dataset \cr}
}
\value{
a dataframe ready to be plotted
}
\description{
Make predictions for the PEM. Series of functions in the PEM MLE.
}
