% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predSurv_present.R
\name{predsu_plot_pintervals}
\alias{predsu_plot_pintervals}
\title{Plot Interval Probabilities}
\usage{
predsu_plot_pintervals(pintervals, lambda_surv = NULL)
}
\description{
Plot Interval Probabilities
}
