% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predSurv_tools.R
\name{predsu_pred}
\alias{predsu_pred}
\title{Predict survival time}
\usage{
predsu_pred(prob_ints, condition_t = 0, option = c("upper", "lower"))
}
\description{
Predict survival time conditioning on censoring time based on interval
probabilities
}
