% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predSurv_tools.R
\name{predsu_pred_all}
\alias{predsu_pred_all}
\title{Predict All}
\usage{
predsu_pred_all(
  dta,
  censor_event = "Dead",
  censor_lfu = "Lost to follow up",
  censor_ana = "Censored at analysis",
  v_enroll = "Time_Enroll",
  v_event = "T_Event",
  v_censor = "Ind_Censor",
  t_ana = 0.75,
  t_enrollment = 1,
  t_next_ana = 2,
  n_to_enroll = 15,
  ...
)
}
\arguments{
\item{dta}{Survival dataset}

\item{censor_event}{Level of censor indicator corresponding to event}

\item{censor_lfu}{Level of censor indicator corresponding to censored due to
lost to follow up}

\item{censor_ana}{Level of censor indicator corresponding to censored due
to analysis time occur first}

\item{v_enroll}{Column name for enrollment time (starting from 0 when the
study started)}

\item{v_event}{Column name for event or censoring time}

\item{v_censor}{Column name for censoring indicator}

\item{t_ana}{Time of the current (interim) analysis (starting from 0 when the study
stared)}

\item{t_enrollment}{Duration of enrollment (starting from 0 when the study
stared). Bigger than t_ana if need to enroll more patients}

\item{t_next_ana}{Time for the next analysis (starting from 0 when the
study stared)}

\item{n_to_enroll}{Number of patients to be enrolled from the time of the
interim analysis to the time enrollment will finish}
}
\value{
A data frame with the following new columns. Pred_T_Event: predicted
    event or censoring time; Pred_Ind_Censor: predicted censoring status;
    Pred_T_Surv: predicted survival time
}
\description{
Predict survival time conditioning on censoring time based on interval
probabilities
}
