% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predSurv_tools.R
\name{predsu_simu}
\alias{predsu_simu}
\title{Simulate Survical Data}
\usage{
predsu_simu(
  n = 300,
  t_ana = 1,
  t_enrollment = 0.25,
  lambda_surv = 1,
  lambda_censor = 0.5
)
}
\arguments{
\item{n}{Sample size}

\item{t_ana}{Time from study starts to the (interim) anlysis}

\item{t_enrollment}{Time from study starts to the enrollment finishes}
}
\value{
A dataframe with enrollment time, time of survival, time of
    censoring, time of event (survival or censoring), and the indicator of
    censoring
}
\description{
Simulate a survival dataset
}
