% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_surv.R
\name{random_forest_plot}
\alias{random_forest_plot}
\title{Plot Random Forest}
\usage{
random_forest_plot(bst)
}
\arguments{
\item{obj}{: survival model object for example a glmnet fit. \cr
coef: coefficient beta is the default. \cr
new_data: a test holdout dataframe default test\cr
train_data : train dataframe default train\cr}
}
\value{
a coxph fit object
}
\description{
Plot Fitted Individual Tree
}
