% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relaxed_lasso.R
\name{rel_shrink}
\alias{rel_shrink}
\title{Relaxed Cox Lasso}
\usage{
rel_shrink(obj, data, time = os_months, status = os_deceased,
  subject = patient_id, table_bool = TRUE, ...)
}
\arguments{
\item{obj}{: survival model object for example a glmnet fit. \cr
train_data : train dataframe default train\cr}
}
\value{
a coxph fit object
}
\description{
This function performs the so called relaxed Lasso for a Cox model.
}
