% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pred.R
\name{roc.plot2}
\alias{roc.plot2}
\title{Plot ROC Curves}
\usage{
roc.plot2(..., x.var = "fpr", y.var = "tpr", x.lab = x.var,
  y.lab = y.var, ident = TRUE)
}
\arguments{
\item{obj}{: survival model object for example a glmnet fit. \cr
coef: coefficient beta is the default. \cr
new_data: a test holdout dataframe default test\cr
train_data : train dataframe default train\cr}
}
\value{
a coxph fit object
}
\description{
Take various ROC curves and plot, a similar function can be found in the caret Tutorial.
}
\author{
Tarj
}
