% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy.R
\name{roc_curve}
\alias{roc_curve}
\title{Calculate Time dependent ROC Curve
There has been a lot of discussing about the usefullness of ROC curves\cr
The time dependent ROC Curves can be used in Survival analysis but only\cr
make sense if all the models to be evaluated are based on a similar link\cr
function. For example all models are Cox Models with different priors,\cr
penalised likelihoods or parameters. However they don't make sense to compare\cr
a tree based method with a Cox Model, or a Weibull against a Cox model \cr
because they structure.}
\usage{
roc_curve(newdata, time = os_months, status = os_status, npi = npi,
  cutoff = 12, plot = F)
}
\arguments{
\item{pred_newdata}{a dataset prediction made by the training set (or CV fold)  \cr
tet_data the hold out dataset \cr
t an optional parameter to integrate the Brier score in a different time, default is max observed time \cr}
}
\value{
pp_bried the posterior predicted integrated brier score
}
\description{
Calculate Time dependent ROC Curve
There has been a lot of discussing about the usefullness of ROC curves\cr
The time dependent ROC Curves can be used in Survival analysis but only\cr
make sense if all the models to be evaluated are based on a similar link\cr
function. For example all models are Cox Models with different priors,\cr
penalised likelihoods or parameters. However they don't make sense to compare\cr
a tree based method with a Cox Model, or a Weibull against a Cox model \cr
because they structure.
}
