% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\name{sim_data}
\alias{sim_data}
\title{Simulate data, function 1 with Weibull distribution.}
\usage{
sim_data(mu, alpha, beta)
}
\arguments{
\item{rate}{rate parameter for the exponential distribution \cr
beta beta regression coefficient\cr}
}
\value{
simulated data
}
\description{
This function simulates a survival dataset given the parameters and covariates.
}
