% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{std_dat}
\alias{std_dat}
\title{Center}
\usage{
std_dat(d)
}
\arguments{
\item{d}{a dataset \cr
... vars to be evaluated \cr
n optional parameter to assure number of rows deleted, default NA}
}
\value{
d a centered dataset
}
\description{
Center
}
