% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stick_vars.R
\name{stick_vars}
\alias{stick_vars}
\title{Stick genes}
\usage{
stick_vars(clidata, genedata)
}
\arguments{
\item{clidata}{: clinical dataset \cr
genedata: genomic dataset \cr}
}
\value{
a full dataset
}
\description{
This function makes easy to stick the gene vars to the clinical vars
}
