% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{survdata}
\alias{survdata}
\title{Data Simulated for a Hypothetic Microarray experiment}
\format{Gene expression matrix with 1000 rows and 40 columns, survival outcome data frame with 40 rows and 2 variables (time, censoring).
\describe{
  \item{surv_months}{Actual survival months observed}
  \item{censoring.status}{censoring indicator 1 observed, 0 censored}
  \item{featurenames}{Hypothetic feature names in the gene expression matrix}
  ...
}}
\usage{
survdata
}
\description{
A dataframe containing the survival outcome and a gene expression matrix refere to data_sim_script to check the algorithm.
}
\keyword{datasets}
