% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predsurv.R
\name{univariate_sel}
\alias{univariate_sel}
\title{Fit and select univariate selection survival model
Get the cost and coefficients for univariate survival analysis.}
\usage{
univariate_sel(d_train, g, x_vars, time_var = "time",
  status_var = "status", threshold = 0.01, correction = "bonferroni",
  ncores = 1L, save_file)
}
\arguments{
\item{d_train}{dataframe}

\item{g}{additional data frame for bootstrapping}

\item{x_vars}{vars to analysis}

\item{time_var}{defults to time}

\item{status_var}{defaults to status}

\item{threshold}{p value}

\item{correction}{correction to apply.}

\item{ncores}{for parallel.}

\item{save_file}{should results be saved on file}
}
\description{
Fit and select univariate selection survival model
Get the cost and coefficients for univariate survival analysis.
}
