% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_surv.R
\name{var_importance}
\alias{var_importance}
\title{Var importance}
\usage{
var_importance(forest = bst, type = "Permutation")
}
\arguments{
\item{obj}{: survival model object for example a glmnet fit. \cr
coef: coefficient beta is the default. \cr
new_data: a test holdout dataframe default test\cr
train_data : train dataframe default train\cr}
}
\value{
a coxph fit object
}
\description{
Plot Fitted Individual Tree
}
