% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_search.R
\name{ontology_search}
\alias{ontology_search}
\title{Examine Ontology}
\usage{
ontology_search(hallmark_list, predictor, coeff = TRUE)
}
\arguments{
\item{hallmark_list:}{list of hallmarks from ontology database \cr
predictor: feature list \cr
coeff : TRUE return hazard coefficient}
}
\value{
predicted linear predictor for each individual in the test set
}
\description{
Check list of Ontology to find genes relevant assessed by cancer biology.
}
