% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predSurv_package.R
\docType{package}
\name{predSurv-package}
\alias{predSurv-package}
\alias{predSurv}
\title{The 'predSurv' package.}
\description{

}
\references{
Chang, M.N. and Shuster, J.J., 1994. Interim analysis for randomized clinical
trials: simulating the predictive distribution of the log-rank test statistic.
Biometrics, pp.827-833.
}
