% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\name{Sim_data}
\alias{Sim_data}
\title{Simulate data, function 1 with Weibull distribution.}
\usage{
Sim_data(N, features, p = 0.02, horizon = 48, init = FALSE, seed = 111,
  CenRate = 1/100, smooth = 10, ...)
}
\arguments{
\item{rate}{rate parameter for the exponential distribution \cr
beta beta regression coefficient\cr}
}
\value{
simulated data
}
\description{
This function simulates a survival dataset given the parameters and covariates.
}
