% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\name{surv_sim_data}
\alias{surv_sim_data}
\title{Simulate data, function 1 with Weibull distribution.}
\usage{
surv_sim_data(mu = -3, alpha = 5, N = 80, features = 100, seed = 111,
  p = 0.25, cens_rate = 1/10)
}
\arguments{
\item{n}{number of individuals default 80\cr
features number of features in the experiment default 1000 \cr}
}
\value{
simulated data
}
\description{
This function simulates a survival dataset given the parameters and covariates.
}
