% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{survdata}
\alias{survdata}
\title{Microarray Dataset}
\format{Gene expression matrix and survival outcome data frame (time, cens).
\describe{
  \item{time}{Actual survival months observed}
  \item{cens}{censoring indicator 1 observed, 0 censored}
  \item{featurenames}{feature names in the gene expression matrix}
  ...
}}
\usage{
survdata
}
\description{
Lung adenocarcinomas were profiled by Beer et al. (2002) using Affymetrix hu6800 microarrays. The data here were normalized from raw .CEL files by RMAExpress (v0.3). The expression matrix contains expression data for 86 patients with 7,129 probe sets.
}
\keyword{datasets}
